switch to matzek's graph implementation statistics table

+++ Madds/src/de/danielsenff/de/madds/util/ByteConverter.java
@@ -2,11 +2,11 @@ package de.danielsenff.de.madds.util;
 
 public class ByteConverter {
 
-	public static float bit2KibiByte(int vale) {
+	public static float bit2KibiByte(long vale) {
 		return (vale / 8) / 1024;
 	}
 	
-	public static float bit2MibiByte(int vale) {
+	public static float bit2MibiByte(long vale) {
 		return (vale / 8) / (1024*1024);
 	}
 }
